Flytt gjenbrukbare funksjoner til felles-fil
